fix:  move files to khd folder
