+++ plugins/accelerated-moe/src/fms_acceleration_moe/framework_plugin_scattermoe.py
@@ -32,14 +32,6 @@ from .utils import (
 # pylint: disable=too-many-instance-attributes
 class ScatterMoEAccelerationPlugin(AccelerationPlugin):
 
-    # NOTE: we cannot do
-    # - require_packages = {"khd"}
-    # this is because the khd fork is not properly packaged as a PyPI project, and so
-    # - "importlib.util.find_spec('khd')" returns, but
-    # - "importlib.metadata.version('kernel-hyperdrive')" does not return
-    # if we decide to extract the kernels, then we do not need to anymore,
-    # https://github.com/foundation-model-stack/fms-acceleration/issues/105
-
     restricted_model_archs = ["GraniteMoeForCausalLM", "MixtralForCausalLM"]
 
     def __init__(self, configurations: Dict[str, Dict]):
